Move CommandSequenceStatus to CommandSequence
They logically belong together.
