+++ test/test_patir_command.rb
@@ -307,80 +307,3 @@ module Patir::Test
     end
   end
 end
-
-class TestCommandSequenceStatus<Minitest::Test
-  def test_new
-    st=Patir::CommandSequenceStatus.new("sequence")
-    assert(!st.running?)
-    assert(!st.success?)
-    assert_equal(:not_executed, st.status)
-    assert_nil(st.step_state(3))
-  end
-  
-  def test_step_equal
-    st=Patir::CommandSequenceStatus.new("sequence")
-    step1=MockCommandObject.new
-    step2=MockCommandWarning.new
-    step3=MockCommandError.new
-    step1.run
-    step1.number=1
-    step2.run
-    step2.number=2
-    step3.run
-    step3.number=3
-    st.step=step1
-    assert_equal(:success, st.status)
-    assert_equal(step1.status, st.step_state(1)[:status])
-    st.step=step2
-    assert_equal(:warning, st.status)
-    st.step=step3
-    assert_equal(:error, st.status)
-    step2.number=1
-    st.step=step2
-    assert_equal(step2.status, st.step_state(1)[:status])
-    assert_equal(:error, st.status)
-    st.step=step1
-    assert_equal(:error, st.status)
-    refute_nil(st.summary)
-  end
-  
-  def test_completed?
-    st=Patir::CommandSequenceStatus.new("sequence")
-    step1=MockCommandObject.new
-    step1.number=1
-    step2=MockCommandWarning.new
-    step2.number=2
-    step3=MockCommandError.new
-    step3.number=3
-    step4=MockCommandObject.new
-    step4.number=4
-    st.step=step1
-    st.step=step2
-    st.step=step3
-    st.step=step4
-    assert(!st.completed?, "should not be complete.")
-    step1.run
-    st.step=step1
-    assert(!st.completed?, "should not be complete.")
-    step2.run
-    st.step=step2
-    assert(!st.completed?, "should not be complete.")
-    step2.strategy=:fail_on_warning
-    st.step=step2
-    assert(st.completed?, "should be complete.")
-    step2.strategy=nil
-    st.step=step2
-    assert(!st.completed?, "should not be complete.")
-    step3.run
-    step3.strategy=:fail_on_error
-    st.step=step3
-    assert(st.completed?, "should be complete.")
-    step3.strategy=nil
-    st.step=step3
-    assert(!st.completed?, "should not be complete.")
-    step4.run
-    st.step=step4
-    assert(st.completed?, "should be complete.")
-    refute_nil(st.summary)
-  end
-end
